+++ ORBITER/src/lib/group_actions/actions/interface_direct_product.cpp
@@ -23,7 +23,7 @@ namespace group_actions {
 
 
 
-static int direct_product_group_element_image_of(action &A, int a,
+static long int direct_product_group_element_image_of(action &A, long int a,
 	void *elt, int verbose_level);
 static void direct_product_group_element_image_of_low_level(action &A,
 	int *input, int *output, void *elt, int verbose_level);
@@ -68,14 +68,14 @@ static void direct_product_group_element_print_latex(action &A,
 static void direct_product_group_element_print_latex_with_print_point_function(
 	action &A,
 	void *elt, std::ostream &ost,
-	void (*point_label)(std::stringstream &sstr, int pt, void *data),
+	void (*point_label)(std::stringstream &sstr, long int pt, void *data),
 	void *point_label_data);
 //static void direct_product_group_element_print_as_permutation(
 //	action &A, void *elt, std::ostream &ost);
 static void direct_product_group_element_print_verbose(action &A,
 	void *elt, std::ostream &ost);
 static void direct_product_group_print_point(action &A,
-	int a, std::ostream &ost);
+	long int a, std::ostream &ost);
 
 
 
@@ -117,13 +117,13 @@ void action_pointer_table::init_function_pointers_direct_product_group()
 
 
 
-static int direct_product_group_element_image_of(action &A,
-		int a, void *elt, int verbose_level)
+static long int direct_product_group_element_image_of(action &A,
+		long int a, void *elt, int verbose_level)
 {
 	int f_v = (verbose_level >= 1);
 	direct_product &P = *A.G.direct_product_group;
 	int *Elt = (int *) elt;
-	int b;
+	long int b;
 
 	if (f_v) {
 		cout << "direct_product_group_element_image_of "
@@ -420,7 +420,7 @@ static void direct_product_group_element_print_latex(
 static void direct_product_group_element_print_latex_with_print_point_function(
 	action &A,
 	void *elt, std::ostream &ost,
-	void (*point_label)(std::stringstream &sstr, int pt, void *data),
+	void (*point_label)(std::stringstream &sstr, long int pt, void *data),
 	void *point_label_data)
 {
 	cout << "direct_product_group_element_print_latex_with_print_point_function "
@@ -468,7 +468,7 @@ static void direct_product_group_element_print_verbose(
 	P.element_print_easy(Elt, ost);
 }
 
-static void direct_product_group_print_point(action &A, int a, ostream &ost)
+static void direct_product_group_print_point(action &A, long int a, ostream &ost)
 {
 	cout << "direct_product_group_print_point "
 			"not yet implemented" << endl;
